refactor: 패키지 분리
- 각 역할에 맞는 패키지로 분리하였습니다.
