+++ src/main/java/lotto/service/RateOfReturnCalculator.java
@@ -1,4 +1,4 @@
-package lotto.model.calculator;
+package lotto.service;
 
 public class RateOfReturnCalculator {
