+++ src/test/java/lotto/service/LottoResultCheckerTest.java
@@ -1,4 +1,4 @@
-package lotto.model.lotto;
+package lotto.service;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -6,6 +6,9 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.Arrays;
 import java.util.List;
+import lotto.model.lotto.Lotto;
+import lotto.model.lotto.LottoResult;
+import lotto.model.lotto.Lottos;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
